Do not set noclose option to `true` during daemonization.
According to official documentation of ruby Process library, if noclose
option is not set to true, it redirects stdin, stdout, stderr to /dev/null.

`nochdir` option is set to true intentionally because we want to run this
daemon in current directory, ruby changes directory to root path otherwise.

For more information please visit:
  https://ruby-doc.org/core-2.3.0/Process.html#method-c-daemon