Fix the design flaw of Hadooken::Consumer
By this change, users will be able to consume messages in two different
executer forms, `:single_thread` or `:multi_thread`.
In `:single_thread` mode, Hadooken will guarantee that all the messages
will be executed in the same order they received,
In `:multi_thread` mode, Hadooken will execute all the messages in a
thread pool which means the execution order is not guaranteed. This
feature must be used wisely!